refactor(perf): split /courses + /learn into server shell + motion island
19th + 20th in the pattern series.
